Refactor NavBlockList.Item to NavBlockListItem
