+++ src/components/NavBlockList.js
@@ -2,8 +2,6 @@ import React from 'react';
 import PropTypes from 'prop-types';
 import styled from 'styled-components';
 import * as CustomPropTypes from '../propTypes';
-import { COLORS } from '../styles/vars';
-import { pxToRem } from '../styles/utils';
 
 export const NavBlockList = props => {
   const { tag: Tag, row, ...rest } = props;
@@ -24,39 +22,7 @@ NavBlockList.propTypes = {
   row: PropTypes.bool,
 };
 
-export const NavBlockListItem = props => {
-  const { tag: Tag, ...rest } = props;
-  return <Tag {...rest} />;
-};
-
-NavBlockListItem.propTypes = {
-  tag: CustomPropTypes.Tag,
-};
-
-NavBlockListItem.defaultProps = {
-  tag: 'li',
-};
-
-NavBlockListItem.displayName = 'NavBlockListItem';
-
-const StyledNavBlockListItem = styled(NavBlockListItem)`
-  background-color: ${COLORS.highlight3};
-  text-transform: uppercase;
-  color: #fff;
-  margin-bottom: ${pxToRem(3)};
-
-  &:last-child {
-    margin-bottom: 0;
-  }
-
-  a {
-    padding: ${pxToRem(12)} ${pxToRem(16)};
-    color: inherit;
-    display: block;
-  }
-`;
-
-const StyledBlockList = styled(NavBlockList)`
+export default styled(NavBlockList)`
   list-style-type: none;
   padding: 0;
   margin: 0;
@@ -73,7 +39,3 @@ const StyledBlockList = styled(NavBlockList)`
     return '';
   }};
 `;
-
-StyledBlockList.Item = StyledNavBlockListItem;
-
-export default StyledBlockList;
